requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiodns==3.1.1
aiofiles==23.2.1
aiogram==2.25.2
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiormq==6.4.2
aiosignal==1.3.1
anyio==4.3.0
APScheduler==3.9.1.post1
apturl==0.5.2
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
Babel==2.9.1
backcall==0.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.12.3
bidict==0.22.1
bleach==3.1.1
blinker==1.7.0
boto3==1.34.2
botocore==1.34.2
Brlapi==0.7.0
bs4==0.0.1
builtwith==1.3.4
CacheControl==0.12.14
cachetools==5.3.2
certifi==2021.5.30
certspy==0.8.0
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.3
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
command-not-found==0.3
commonmark==0.9.1
confluent-kafka==2.3.0
cryptography==41.0.7
cupshelpers==1.0
cycler==0.12.1
dask==2022.10.2
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
diff-match-patch==20230430
distro==1.8.0
distro-info==0.23+ubuntu1.1
Django==4.2.14
django-cors-headers==4.4.0
django-import-export==4.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
docopt==0.6.2
dsinternals==1.2.4
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fasteners==0.14.1
fbmessenger==6.0.0
fire==0.5.0
flash==1.0.3
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.46.0
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
gast==0.4.0
google==3.0.0
google-auth==2.25.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googlesearch-python==1.2.3
gorilla-cli==0.0.10
gpg==1.13.1
greenlet==3.0.2
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
halo==0.0.31
html5lib==1.0.1
httpcore==0.16.3
httplib2==0.14.0
httptools==0.6.1
httpx==0.23.1
humanfriendly==10.0
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==5.2.0
ipython==7.13.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
isort==5.12.0
itsdangerous==2.1.2
jax==0.4.13
jedi==0.15.2
Jinja2==3.0.1
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.2
jsonschema==4.17.3
jupyter-client==6.1.2
jupyter-core==4.6.3
keras==2.12.0
keyring==18.0.1
kiwisolver==1.4.5
knockpy @ git+https://github.com/guelfoweb/knock.git@8c4d7b50c3662d5d6a423a0f6c2f2f5dec841732
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
ldap3==2.9.1
libclang==16.0.6
locket==1.0.0
lockfile==0.12.2
log-symbols==0.0.14
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
magic-filter==1.0.12
Mako==1.1.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.5.3
mattermostwrapper==2.2
mccabe==0.7.0
mdurl==0.1.2
mistune==0.8.4
ml-dtypes==0.2.0
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.7
multidict==5.2.0
mysqlclient==2.2.4
nbconvert==5.6.1
nbformat==5.0.4
neotermcolor==2.0.10
netifaces==0.10.4
networkx==2.6.3
notebook==6.0.3
npx==0.1.6
numpy==1.23.5
oauthlib==3.2.2
olefile==0.46
openpyxl==3.1.2
opt-einsum==3.3.0
oxdork==3.2.0
packaging==24.1
pamqp==3.2.1
pandas==2.0.3
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.2
partd==1.4.1
pefile==2023.2.7
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
pluggy==1.3.0
portalocker==2.8.2
prometheus-client==0.7.1
prompt-toolkit==3.0.28
protobuf==4.23.3
psutil==5.5.1
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.16.2
pycares==4.4.0
pycodestyle==2.11.1
pycparser==2.21
pycrypto==2.6.1
pycryptodomex==3.19.0
pycups==1.9.73
pydantic==1.10.9
pydot==1.4.2
pyflakes==3.1.0
Pygments==2.17.2
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==2.8.0
pykwalify==1.8.0
pymacaroons==0.13.0
pymongo==4.3.3
pymssql==2.2.10
PyNaCl==1.3.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
pypng==0.20220715.0
PyQt5==5.14.1
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyRFC3339==1.1
pyrsistent==0.20.0
pytesseract==0.3.10
python-apt==2.0.1+ubuntu0.20.4.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
python-engineio==4.8.0
python-magic==0.4.16
python-nmap==0.7.1
python-pam==1.8.4
python-socketio==5.10.0
python-whois==0.8.0
python-xapp==1.8.1
python-xlib==0.23
pytrends==4.9.2
pytz==2022.7.1
pyxdg==0.26
PyYAML==6.0.1
pyzmq==18.1.1
qrcode==7.4.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
questionary==1.10.0
randomname==0.1.5
rasa==3.6.15
rasa-sdk==3.6.2
redis==4.6.0
regex==2022.10.31
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
requests-unixsocket==0.2.0
rfc3986==1.5.0
rich==12.6.0
rocketchat-API==1.30.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
s3transfer==0.9.0
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scapy==2.4.3
scikit-learn==1.1.3
scipy==1.10.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sentry-sdk==1.14.0
setproctitle==1.1.10
shell_gpt==0.9.4
simple-websocket==1.0.0
simplejson==3.16.0
sip==4.19.21
six==1.16.0
sklearn-crfsuite==0.3.6
slack-sdk==3.26.1
sniffio==1.3.1
soupsieve==1.9.5
spinners==0.0.24
SQLAlchemy==1.4.50
sqlparse==0.5.0
ssh-import-id==5.10
structlog==23.2.0
structlog-sentry==2.0.3
systemd-python==234
tablib==3.5.0
tabulate==0.9.0
tarsafe==0.0.4
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-text==2.12.0
termcolor==2.3.0
terminado==0.8.2
terminaltables==3.1.10
testpath==0.4.4
threadpoolctl==3.2.0
tinycss==0.4
tk==0.1.0
tkdesigner==1.0.7
tomli==2.0.1
toolz==0.12.0
tornado==5.1.1
tqdm==4.66.2
traitlets==4.3.3
twilio==8.2.2
typer==0.9.0
typing-utils==0.1.0
typing_extensions==4.8.0
tzdata==2024.1
tzlocal==5.2
ubuntu-cleaner==1.1.4
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
ujson==5.9.0
unattended-upgrades==0.1
urllib3==1.26.6
usb-creator==0.3.7
uvloop==0.19.0
vboxapi==1.0
wadllib==1.3.3
wcwidth==0.2.12
webencodings==0.5.1
webexteamssdk==1.6.1
WebSecProbe==0.0.11
websockets==10.4
websploit==4.0.4
Werkzeug==3.0.1
widgetsnbextension==2.0.0
wrapt==1.14.1
wsproto==1.2.0
xkit==0.0.0
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.17.0
